Add RefreshQuery

+++ job_test.go
@@ -48,7 +48,7 @@ func Test_GetJob_OK(t *testing.T) {
 			ID:            "623b290a-7fd9-4ea6-a2a6-96f9c9101f51",
 			QueryResultID: 1,
 			Status:        3,
-			UpdatedAt:     0,
+			UpdatedAt:     float64(0),
 		},
 	}, res)
 }
